Refactor remote debugging commands into common library, see Microsoft/vscode-azuretools#522 (#1045)
