+++ extension.bundle.ts
@@ -24,4 +24,4 @@ export * from 'vscode-azureextensionui';
 export { SubscriptionTreeItem } from './src/explorer/SubscriptionTreeItem';
 export { AzureAccountTreeItem } from './src/explorer/AzureAccountTreeItem';
 export { getRandomHexString } from './src/utils/randomUtils';
-export { checkForRemoteDebugSupport } from './src/commands/remoteDebug/remoteDebugCommon';
+export { checkForRemoteDebugSupport } from './src/commands/remoteDebug/checkForRemoteDebugSupport';
