+++ src/extension.ts
@@ -27,7 +27,6 @@ import { viewCommitInGitHub } from './commands/deployments/viewCommitInGitHub';
 import { viewDeploymentLogs } from './commands/deployments/viewDeploymentLogs';
 import { enableFileLogging } from './commands/enableFileLogging';
 import { startRemoteDebug } from './commands/remoteDebug/startRemoteDebug';
-import { stopRemoteDebug } from './commands/remoteDebug/stopRemoteDebug';
 import { showFile } from './commands/showFile';
 import { startSsh } from './commands/startSsh';
 import { startStreamingLogs } from './commands/startStreamingLogs';
@@ -268,7 +267,6 @@ export async function activateInternal(
         registerCommand('appService.LogPoints.OpenScript', openScript);
 
         registerCommand('appService.StartRemoteDebug', startRemoteDebug);
-        registerCommand('appService.StopRemoteDebug', stopRemoteDebug);
         registerCommand('appService.StartSsh', startSsh);
 
         registerCommand('appService.showFile', async (_actionContext: IActionContext, node: FileTreeItem) => { await showFile(node, fileEditor); }, 500);
